backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
prophet==1.1.5
pandas==2.1.4
numpy==1.26.2
firebase-admin==6.4.0
python-dotenv==1.0.0
reportlab==4.0.8
gunicorn==21.2.0
pyarrow==14.0.1
holidays==0.38
